+++ integration/helpers/vite.ts
@@ -30,6 +30,7 @@ export const reactRouterConfig = ({
   appDirectory,
   splitRouteModules,
   viteEnvironmentApi,
+  middleware,
 }: {
   ssr?: boolean;
   basename?: string;
@@ -39,6 +40,7 @@ export const reactRouterConfig = ({
     Config["future"]
   >["unstable_splitRouteModules"];
   viteEnvironmentApi?: boolean;
+  middleware?: boolean;
 }) => {
   let config: Config = {
     ssr,
@@ -48,6 +50,7 @@ export const reactRouterConfig = ({
     future: {
       unstable_splitRouteModules: splitRouteModules,
       unstable_viteEnvironmentApi: viteEnvironmentApi,
+      unstable_middleware: middleware,
     },
   };
 
@@ -99,6 +102,7 @@ export const viteConfig = {
 
 export const EXPRESS_SERVER = (args: {
   port: number;
+  base?: string;
   loadContext?: Record<string, unknown>;
 }) =>
   String.raw`
@@ -143,7 +147,8 @@ export const EXPRESS_SERVER = (args: {
 export type TemplateName =
   | "vite-5-template"
   | "vite-6-template"
-  | "vite-cloudflare-template";
+  | "cloudflare-dev-proxy-template"
+  | "vite-plugin-cloudflare-template";
 
 export const viteMajorTemplates = [
   { templateName: "vite-5-template", templateDisplayName: "Vite 5" },
@@ -366,7 +371,7 @@ export const test = base.extend<Fixtures>({
       let port = await getPort();
       let cwd = await createProject(
         await files({ port }),
-        "vite-cloudflare-template"
+        "cloudflare-dev-proxy-template"
       );
       let { status } = build({ cwd });
       expect(status).toBe(0);